pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zenith-ai-cli"
version = "0.1.0"
description = "🏔️ ZENITH: The Peak of Terminal Productivity"
authors = [
    { name = "Zenith Team" }
]
dependencies = [
    "typer[all]",
    "rich",
    "httpx",
    "python-dotenv",
    "pydantic",
    "questionary",
    "google-generativeai"
]
requires-python = ">=3.10"

[project.scripts]
zenith = "zenith.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["zenith*"]
